Add World Component

+++ include/Collision.h
@@ -22,7 +22,7 @@ struct Ray {
 	Ray() = default;
 	Ray(glm::vec3 start, glm::vec3 dir);
 
-	std::vector<glm::vec3> getAffectedBlocks(float length) const;
+	std::vector<glm::ivec3> getAffectedBlocks(float length) const;
 };
 
 typedef std::pair<glm::vec3, glm::vec3> Line;
